requirements.txt
asgiref==3.4.1
Django==3.2.7
django-location-field==2.1.0
django-redis==5.0.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
zope.event==4.5.0
zope.interface==5.4.0
